add IPFSTimeInfo components.
