+++ components/DataViewer.tsx
@@ -10,12 +10,12 @@ import renderHTML from 'react-render-html';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import CustomerValidations from './CustomerValidations';
-import { getArweaveBlockByHash, getArweaveTxnStatusByHash } from '../services/arweave';
 import { ShowItem } from './PageElements';
 import DataSourceContext from '../utils/dataSource';
 import testPayloads from '../utils/testPayloads.json';
 import { useMetadata } from '../services/metadata';
 import ArweaveTxnStatus from './ArweaveTxnStatus';
+import IPFSTimeInfo from './IPFSTimeInfo';
 import { getCidTimeInfo, IPFSCidTimeInfoMappingContractAddress, chainInfo, getTxnHashByCidAndBlockNumberFromRPC } from '../services/IPFSCidTimeInfoMapping';
 
 const md = require('markdown-it')().use(require('markdown-it-plantuml'));
@@ -103,28 +103,28 @@ function DataViewer<TMetadataType>(props: IDataViewerProps) {
   //   setBlockTimestamp(timestamp * 1000);
   // }, [options.id, setBlockNumber, setBlockTimestamp]);
 
-  const getIPFSTimeInfo = useCallback(async () => {
-    if (!options.id) return;
-    try {
-      const { timestamp, blockNumber } = await getCidTimeInfo(options.id);
-      setBlockNumber(Number(blockNumber));
-      setBlockTimestamp(Number(timestamp) * 1000);
-      const hash = await getTxnHashByCidAndBlockNumberFromRPC(options.id, Number(blockNumber));
-      console.log(hash);
-      setRemark({
-        hash: {
-          title: "Txn hash",
-          content: hash,
-          url: `${chainInfo.scan}tx/${hash}`,
-          urlTitle: `${chainInfo.name} explorer`,
-          type: 'time'
-        }
-      });
-    } catch (error) {
-      console.log(`get IPFS timeinfo failure.`);
-      // console.log(error);
-    }
-  }, [options.id, setBlockNumber, setBlockTimestamp, setRemark]);
+  // const getIPFSTimeInfo = useCallback(async () => {
+  //   if (!options.id) return;
+  //   try {
+  //     const { timestamp, blockNumber } = await getCidTimeInfo(options.id);
+  //     setBlockNumber(Number(blockNumber));
+  //     setBlockTimestamp(Number(timestamp) * 1000);
+  //     const hash = await getTxnHashByCidAndBlockNumberFromRPC(options.id, Number(blockNumber));
+  //     console.log(hash);
+  //     setRemark({
+  //       hash: {
+  //         title: "Txn hash",
+  //         content: hash,
+  //         url: `${chainInfo.scan}tx/${hash}`,
+  //         urlTitle: `${chainInfo.name} explorer`,
+  //         type: 'time'
+  //       }
+  //     });
+  //   } catch (error) {
+  //     console.log(`get IPFS timeinfo failure.`);
+  //     // console.log(error);
+  //   }
+  // }, [options.id, setBlockNumber, setBlockTimestamp, setRemark]);
 
   useEffect(() => {
     (id && dataSource) ? getMetadata() : null;
@@ -142,7 +142,7 @@ function DataViewer<TMetadataType>(props: IDataViewerProps) {
       }
       if (options.platform === 'ipfs') {
         // TODO SWITCH TO SWR
-        getIPFSTimeInfo();
+        // getIPFSTimeInfo();
       }
     } catch (error) {
       if (metadataResult.isError) {
@@ -154,7 +154,7 @@ function DataViewer<TMetadataType>(props: IDataViewerProps) {
     }
 
 
-  }, [options, getIPFSTimeInfo, metadataResult.isError]);
+  }, [options, metadataResult.isError]);
 
   useEffect(() => {
     console.log(`Metadata version: ${metadata['@version']}`);
@@ -266,37 +266,7 @@ function DataViewer<TMetadataType>(props: IDataViewerProps) {
                     options.platform == 'arweave' && !(metadataResult.isError) ? <ArweaveTxnStatus hash={options.id} /> : <></>
                   }
                   {
-                    options.platform == 'ipfs' && blockNumber > 0 ? <div className="mt-8">
-                      <h2 className="font-thin text-2xl text-purple-700">Time information</h2>
-                      <ShowItem
-                        title="Block"
-                        content={blockNumber.toString()}
-                      />
-                      <ShowItem
-                        title="Timestamp"
-                        content={new Date(blockTimestamp).toLocaleString()}
-                      />
-                      <ShowItem
-                        title="Contract"
-                        content={IPFSCidTimeInfoMappingContractAddress}
-                        url={`${chainInfo.scan}address/${IPFSCidTimeInfoMappingContractAddress}`}
-                        urlTitle={`${chainInfo.name} explorer`}
-                      />
-                      {
-                        remark && (remark as any).hash ? <ShowItem
-                          title="Txn hash"
-                          content={(remark as any).hash.content}
-                          url={(remark as any).hash.url}
-                          urlTitle={(remark as any).hash.urlTitle}
-                        /> : <></>
-                      }
-                    </div> : (options.platform == 'ipfs' && blockNumber === 0 ? <div className="mt-8">
-                      <h2 className="font-thin text-2xl text-purple-700">Time information</h2>
-                      <ShowItem
-                        title="Time infomation not found."
-                        content={'No deposited information found.'}
-                      />
-                    </div> : options.platform == 'ipfs' ? <div className="text-xs font-thin text-purple-500 animate-pulse mt-4">Query CID time infomation...</div> : <></>)
+                    options.platform == 'ipfs' ? <IPFSTimeInfo ipfsHash={options.id} platform={options.platform} /> : <></>
                   }
 
                   {
